+++ changelog.md
@@ -5,6 +5,8 @@
 - added save button
 - added load button
 - fixed bug with the loading function
+- added shop with ability to upgrade the cash
+- added button to return to the main screen (the one with the machine)
 
 ## Version 0.1:
 - added the machine as a complex object